required for new glib
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@31035 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/rules/asterisk.mk
@@ -1,51 +1,3 @@
-util-linux-configure:
-	make -C util-linux clean
-	cd util-linux && ./configure --host=$(ARCH)-linux-uclibc --prefix=/usr --libdir=/usr/lib CFLAGS="$(COPTS) $(MIPS16_OPT) -DNEED_PRINTF" PKG_CONFIG="/tmp" NCURSES_CFLAGS="-I$(TOP)/ncurses/include" NCURSES_LIBS="-L$(TOP)/ncurses/lib -lncurses" \
-	--disable-rpath \
-	--enable-new-mount	\
-	--disable-tls		\
-	--disable-sulogin	\
-	--without-python	\
-	--without-udev		\
-	--with-ncurses
-	make -C util-linux
-
-util-linux-clean:
-	make -C util-linux clean
-
-util-linux:
-	make -C util-linux clean
-	make -C util-linux
-
-util-linux-install:
-	make -C util-linux clean
-	make -C util-linux
-	make -C util-linux install DESTDIR=$(INSTALLDIR)/util-linux
-	rm -f $(INSTALLDIR)/util-linux/usr/lib/libuuid.a
-	rm -f $(INSTALLDIR)/util-linux/usr/lib/libblkid.so*
-	rm -f $(TOP)/util-linux/.libs/libuuid.a
-	rm -f $(TOP)/util-linux/.libs/libblkid.a
-ifneq ($(CONFIG_ASTERISK),y)
-	rm -rf $(INSTALLDIR)/util-linux/usr/sbin
-	rm -rf $(INSTALLDIR)/util-linux/usr/bin
-	rm -rf $(INSTALLDIR)/util-linux/usr/share
-	rm -rf $(INSTALLDIR)/util-linux/usr/include
-	rm -rf $(INSTALLDIR)/util-linux/usr/lib/pkgconfig
-	rm -f $(INSTALLDIR)/util-linux/usr/lib/libmount*
-	rm -f $(INSTALLDIR)/util-linux/usr/lib/libfdisk*
-	rm -f $(INSTALLDIR)/util-linux/usr/lib/libsmartcols*
-	rm -f $(INSTALLDIR)/util-linux/usr/lib/libblkid.a
-	rm -f $(INSTALLDIR)/util-linux/usr/lib/libblkid.la
-	rm -f $(INSTALLDIR)/util-linux/usr/lib/libuuid.a
-	rm -f $(INSTALLDIR)/util-linux/usr/lib/libuuid.la
-	rm -f $(INSTALLDIR)/util-linux/usr/lib/libblkid.so*
-	rm -rf $(INSTALLDIR)/util-linux/bin
-	rm -rf $(INSTALLDIR)/util-linux/sbin
-	rm -f $(INSTALLDIR)/util-linux/lib/libmount.so*
-	rm -f $(INSTALLDIR)/util-linux/lib/libfdisk.so*
-	rm -f $(INSTALLDIR)/util-linux/lib/libsmartcols.so*
-endif
-
 asterisk-configure: util-linux-configure util-linux-install jansson
 	rm -f asterisk/menuselect.makeopts && \
 	cd asterisk && ./configure --host=$(ARCH)-linux-uclibc \
@@ -230,10 +182,20 @@ asterisk-install:
 	rm -f $(INSTALLDIR)/util-linux/usr/lib/libblkid.la
 	rm -f $(INSTALLDIR)/util-linux/usr/lib/libuuid.a
 	rm -f $(INSTALLDIR)/util-linux/usr/lib/libuuid.la
-	rm -f $(INSTALLDIR)/util-linux/usr/lib/libblkid.so*
 	rm -rf $(INSTALLDIR)/util-linux/bin
 	rm -rf $(INSTALLDIR)/util-linux/sbin
+ifneq ($(CONFIG_ASTERISK),y)
+ifneq ($(CONFIG_ZABBIX),y)
+ifneq ($(CONFIG_MC),y)
+ifneq ($(CONFIG_LIBQMI),y)
+ifneq ($(CONFIG_WEBSERVER),y)
+	rm -f $(INSTALLDIR)/util-linux/usr/lib/libblkid.so*
 	rm -f $(INSTALLDIR)/util-linux/lib/libmount.so*
+endif
+endif
+endif
+endif
+endif
 	rm -f $(INSTALLDIR)/util-linux/lib/libfdisk.so*
 	rm -f $(INSTALLDIR)/util-linux/lib/libsmartcols.so*
